Remove record-based EEPROM emulation
